+++ CHANGES.md
@@ -5,6 +5,7 @@
 ##### Java Client
 
 * Products can be found by using the query API.
+* Categories can be obtained as sorted list for one level.
 
 ## v0.48.0